Align GPU-active clipboard export in the frametime view
The frametime view's copy functions paired frametimes and GPU active
times by index. Validity filtering can drop different samples from
each series, so pair by timestamp intersection like the FPS view.
The shared helper moves to GraphDataContextBase.
